Update faq.mdx
I have fixed a **grammatical** mistake in this document. 
In the first question titled as **I am new to GitHub and Open Source. Where should I start?**, the answer has a grammatical typo. 

The typo is -:

`includes tips on contributing to open-source projects`

The fixed sentence is -:

`include tips on contributing to open-source projects`